+++ intents/language/agent_language.py
@@ -1,6 +1,7 @@
 import os
 import sys
 import logging
+import inspect
 from typing import List, Type
 
 from intents.language_codes import LanguageCode, LANGUAGE_CODES, FALLBACK_LANGUAGE
@@ -8,14 +9,12 @@ from intents.language_codes import LanguageCode, LANGUAGE_CODES, FALLBACK_LANGUA
 logger = logging.getLogger(__name__)
 
 def agent_language_folder(agent_cls: Type["intents.model.agent.Agent"]) -> str:
-    main_agent_package_name = agent_cls.__module__.split('.')[0]
-    main_agent_package = sys.modules[main_agent_package_name]
-    if '__path__' not in main_agent_package.__dict__:
-        # TODO: try workdir or something...
+    try:
+        agent_folder = os.path.dirname(inspect.getfile(agent_cls))
+    except TypeError:
         logger.warning("Agent %s doesn't seem to be defined within a package. Language data will not be loaded.", agent_cls)
         return [], []
 
-    agent_folder = main_agent_package.__path__[0]
     language_folder = os.path.join(agent_folder, 'language')
     if not os.path.isdir(language_folder):
         raise ValueError(f"No language folder found for agent {agent_cls} (expected: {language_folder})")